requirements.txt
google-generativeai==0.8.3
ollama==0.3.3
streamlit==1.39.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
beautifulsoup4==4.12.3
fake-useragent==1.5.1
plotly==5.24.1
email-validator==2.2.0
